Refactor Login to use API & fix bug
Login error "no account found" wasn't working
