+++ src/app/libs/logIn.js
@@ -1,21 +1,15 @@
 import dummyData from './dummyData.json';
+import getAllElder from './api/getElder';
+
 /**
+ * Do an API call to get all the elder records
  * Search the database for a record's phone number that match the user enter number. 
  * If found, update localstorage (app state) with the single data object and send a true boolean.
  * Else, send a false boolean.
  */
-const getElder = (phoneNumber) => {
-    let parseData;
-
-    // Check if the test data is already saved. If not, save it. If so, use it.
-    let previousTestData = localStorage.getItem("Elder-test-data");
-    if (previousTestData === null) {
-      localStorage.setItem("Elder-test-data", JSON.stringify(dummyData)); // TODO: Swap with a real database call
-      parseData = dummyData;
-    } else {
-        let previousSavedData = localStorage.getItem("Elder-test-data"); // TODO: Swap with a real database call
-        parseData = JSON.parse(previousSavedData)
-    }
+const getElder = async (phoneNumber) => {
+    let parseData = await getAllElder();
+    let enableLogin = false;
 
     let foundIndex = parseData.findIndex((data) => {
         return data.phoneNumber === phoneNumber;
@@ -23,10 +17,12 @@ const getElder = (phoneNumber) => {
 
     if(foundIndex !== -1) {
         localStorage.setItem("Elder-data", JSON.stringify(parseData[foundIndex]));
-        return true;
+        enableLogin =  true;
     }else {
-        return false;
+        enableLogin = false;
     }
+   
+    return enableLogin
 }
 
 export {getElder}
